# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Avito PVZ Service 2025",
        "title": "PVZ Service",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/receptions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новую приемку товаров для указанного ПВЗ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "receptions"
                ],
                "summary": "Создать приемку",
                "operationId": "create-reception",
                "parameters": [
                    {
                        "description": "Информация о создаваемой приемке",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateReceptionRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Приемка успешно создана",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateReceptionResponseDto"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные / ПВЗ не найден / ПВЗ уже имеет незакрытую приемку",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateReceptionRequestDto": {
            "description": "Информация о приемке при ее создании",
            "type": "object",
            "properties": {
                "pvzId": {
                    "description": "Идентификатор ПВЗ",
                    "type": "string"
                }
            }
        },
        "dto.CreateReceptionResponseDto": {
            "description": "Информация о приемке при ее создании",
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "Дата и время приемки",
                    "type": "string"
                },
                "id": {
                    "description": "Идентификатор приемки",
                    "type": "string"
                },
                "pvzId": {
                    "description": "Идентификатор ПВЗ",
                    "type": "string"
                },
                "status": {
                    "description": "Статус приемки",
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorDto": {
            "description": "Информация об ошибке (DTO)",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Текст ошибки",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Authorization check",
            "type": "apiKey",
            "name": "auth-x",
            "in": "header"
        }
    }
}